+++ main.ts
@@ -4,6 +4,7 @@ interface MetadataHiderSettings {
 	hideEmptyEntryInSiderbar: boolean;
 	propertiesVisible: string;
 	propertiesInvisible: string;
+	propertiesInvisibleAlways: string;
 	propertyHideAll: string;
 }
 
@@ -11,6 +12,7 @@ const DEFAULT_SETTINGS: MetadataHiderSettings = {
 	hideEmptyEntryInSiderbar: false,
 	propertiesVisible: "",
 	propertiesInvisible: "",
+	propertiesInvisibleAlways: "",
 	propertyHideAll: "hide",
 }
 
@@ -18,15 +20,63 @@ export default class MetadataHider extends Plugin {
 	settings: MetadataHiderSettings;
 	styleTag: HTMLStyleElement;
 
+	isMetadataFocused: boolean;
+
 
 	async onload() {
+		console.log("metadata hider onload")
 		await this.loadSettings();
 		this.addSettingTab(new MetadataHiderSettingTab(this.app, this));
 		this.updateCSS();
+
+		this.registerDomEvent(document, 'focusin', (evt: MouseEvent) => {
+			console.log('focusin', evt);
+			const target = evt.target;
+			const metadataElement = document.querySelector('.workspace-leaf.mod-active .metadata-container');
+			if (metadataElement === null) { return; }
+			if (metadataElement?.contains(target as Node)) {
+				console.log(target)
+				metadataElement.classList.add('is-active');
+				this.isMetadataFocused = true;
+				// @ts-ignore
+				if (target?.classList?.contains("metadata-add-button")) {
+					const clickEvent = new MouseEvent('click', {
+						bubbles: true,
+						cancelable: true,
+						view: window
+					});
+					target.dispatchEvent(clickEvent);
+				}
+			} else if (this.isMetadataFocused) {
+				this.isMetadataFocused = false;
+				metadataElement.classList.remove('is-active');
+			}
+		});
+		this.registerDomEvent(document, 'focusout', (evt: MouseEvent) => {
+			console.log('focusout', evt);
+			const target = evt.target;
+			const metadataElement = document.querySelector('.workspace-leaf.mod-active .metadata-container');
+			if (metadataElement?.contains(target as Node)) {
+				this.isMetadataFocused = false;
+				setTimeout(() => {
+					if (!this.isMetadataFocused) {
+						metadataElement.classList.remove('is-active');
+					}
+				}, 100);
+			}
+
+		});
 	}
 
 	onunload() {
-		this.styleTag.remove();
+		const parentElement = this.styleTag.parentElement;
+		if (parentElement) {
+			parentElement.removeChild(this.styleTag);
+			console.log('Style element removed successfully.');
+		} else {
+			console.error('Parent element not found.');
+		}
+		// this.styleTag.remove();
 	}
 
 	debounceUpdateCSS = debounce(this.updateCSS, 1000, true);
@@ -42,7 +92,7 @@ export default class MetadataHider extends Plugin {
 		}
 
 		headElement.appendChild(this.styleTag);
-		this.styleTag.innerText = genSnippetCSS(this);
+		this.styleTag.innerText = genAllCSS(this);
 	}
 
 
@@ -63,48 +113,32 @@ function genCSS(properties: string, cssPrefix: string, cssSuffix: string, parent
 	}
 	let body: string[] = [];
 	parentSelector = parentSelector ? parentSelector + " " : "";
+	properties = properties.replace(/\n/g, "");
 	for (let property of properties.split(',')) {
-		body.push(`${parentSelector} .metadata-container > .metadata-content > .metadata-properties > .metadata-property[data-property-key="${property.trim()}"]`);
+		body.push(`${parentSelector}.metadata-container > .metadata-content > .metadata-properties > .metadata-property[data-property-key="${property.trim()}"]`);
 	}
 	const sep = " ";
 	return cssPrefix + sep + body.join(',' + sep) + sep + cssSuffix + sep + sep;
 }
 
-function genSnippetCSS(plugin: MetadataHider): string {
+function genAllCSS(plugin: MetadataHider): string {
 	const s = plugin.settings;
-	const mod_root = s.hideEmptyEntryInSiderbar ? "" : ".mod-root ";
+	const mod_root = s.hideEmptyEntryInSiderbar ? "" : ".mod-root  ";
 	const content: string[] = [
+		`.metadata-container.is-active .metadata-property { display: flex !important; }`,
 		/* * Hide the metadata that is empty */
 		`${mod_root}.metadata-property:has(.metadata-property-value .mod-truncate:empty),`,
-		`${mod_root}.metadata-property:has(`,
-		`		.metadata-property-value`,
-		`			input.metadata-input[type="number"]:placeholder-shown`,
-		`	),`,
+		`${mod_root}.metadata-property:has(.metadata-property-value input.metadata-input[type="number"]:placeholder-shown),`,
 		`${mod_root}.metadata-property[data-property-type="text"]:has(input[type="date"]),`,
-		`${mod_root}.metadata-property:has(`,
-		`		.metadata-property-value`,
-		`			.multi-select-container`,
-		`			> .multi-select-input:first-child`,
-		`	) {`,
+		`${mod_root}.metadata-property:has(.metadata-property-value .multi-select-container > .multi-select-input:first-child) {`,
 		`	display: none;`,
 		`}`,
-		/* * visualize the last metadata property.`
-		 * Make sure to enable Button: add document attribution */
-		`${mod_root}div.view-content`,
-		`	div.metadata-content`,
-		`	> div.metadata-properties`,
-		`	> div.metadata-property:last-child {`,
-		`	display: flex;`,
-		`}`,
-		``,
 	];
 
 
 	if (s.propertyHideAll.trim()) {
 		content.push([
-			`${mod_root}.metadata-container:has(`,
-			`	.metadata-property[data-property-key="${s.propertyHideAll.trim()}"] input[type="checkbox"]:checked`,
-			`  ) {`,
+			`${mod_root}.metadata-container:has(.metadata-property[data-property-key="${s.propertyHideAll.trim()}"] input[type="checkbox"]:checked) {`,
 			`  display: none;`,
 			`}`,
 			``,
@@ -154,8 +188,8 @@ class MetadataHiderSettingTab extends PluginSettingTab {
 			});
 
 		new Setting(containerEl)
-			.setName({ en: "Always display metadata properties", zh: "永远显示的文档属性（元数据）", "zh-TW": "永遠顯示的文件屬性（元數據）" }[lang] as string)
-			.setDesc({ en: "separated by comma (`,`)", zh: "英文逗号分隔（`,`）。如：“tags, aliases”", "zh-TW": "以逗號分隔（`,`）" }[lang] as string)
+			.setName({ en: "Metadata properties that keep displaying", zh: "永远显示的文档属性（元数据）", "zh-TW": "永遠顯示的文件屬性（元數據）" }[lang] as string)
+			.setDesc({ en: "Metadata properties will always display even if their value are empty. Metadata property keys are separated by comma (`,`)", zh: "英文逗号分隔（`,`）。如：“tags, aliases”", "zh-TW": "以逗號分隔（`,`）" }[lang] as string)
 			.addTextArea((text) =>
 				text
 					.setValue(this.plugin.settings.propertiesVisible)
@@ -166,8 +200,8 @@ class MetadataHiderSettingTab extends PluginSettingTab {
 					})
 			);
 		new Setting(containerEl)
-			.setName({ en: "Always hide metadata properties", zh: "永远隐藏的文档属性（元数据）", "zh-TW": "永遠隱藏的文件屬性（元數據）" }[lang] as string)
-			.setDesc({ en: "separated by comma (`,`)", zh: "英文逗号分隔（`,`）。如：“tags, aliases”", "zh-TW": "以逗號分隔（`,`）" }[lang] as string)
+			.setName({ en: "Metadata properties to hide", zh: "隐藏的文档属性（元数据）", "zh-TW": "永遠隱藏的文件屬性（元數據）" }[lang] as string)
+			.setDesc({ en: "Metadata properties will always hide even if their value are not empty, but will display when the metadata properties table is focused. Metadata property keys are separated by comma (`,`)", zh: "英文逗号分隔（`,`）。如：“tags, aliases”", "zh-TW": "以逗號分隔（`,`）" }[lang] as string)
 			.addTextArea((text) =>
 				text
 					.setValue(this.plugin.settings.propertiesInvisible)
@@ -178,8 +212,20 @@ class MetadataHiderSettingTab extends PluginSettingTab {
 					})
 			);
 		new Setting(containerEl)
-			.setName({ en: "Hide the whole metadata properties table", zh: "隐藏整个文档属性（元数据）表格", "zh-TW": "隱藏整個文檔屬性（元數據）表格" }[lang] as string)
-			.setDesc({ en: `when its value is true`, zh: `当该属性值为真时`, "zh-TW": `當該屬性值為真時` }[lang] as string)
+			.setName({ en: "Metadata properties always to hide", zh: "永远隐藏的文档属性（元数据）", "zh-TW": "永遠隱藏的文件屬性（元數據）" }[lang] as string)
+			.setDesc({ en: "Metadata properties will always hide even if their value are not empty or the metadata properties table is focused. Metadata property keys are separated by comma (`,`)", zh: "英文逗号分隔（`,`）。如：“tags, aliases”", "zh-TW": "以逗號分隔（`,`）" }[lang] as string)
+			.addTextArea((text) =>
+				text
+					.setValue(this.plugin.settings.propertiesInvisibleAlways)
+					.onChange(async (value) => {
+						this.plugin.settings.propertiesInvisibleAlways = value;
+						await this.plugin.saveSettings();
+						this.plugin.debounceUpdateCSS();
+					})
+			);
+		new Setting(containerEl)
+			.setName({ en: "Key to hide the whole metadata properties table", zh: "隐藏整个文档属性（元数据）表格", "zh-TW": "隱藏整個文檔屬性（元數據）表格" }[lang] as string)
+			.setDesc({ en: `when its value is true, the whole metadata properties table will be hidden`, zh: `当该属性值为真时`, "zh-TW": `當該屬性值為真時` }[lang] as string)
 			.addSearch((cb) => {
 				cb.setPlaceholder({ en: "entry name", zh: "文档属性名称", "zh-TW": "文件屬性名稱", }[lang] as string)
 					.setValue(this.plugin.settings.propertyHideAll)
@@ -193,11 +239,11 @@ class MetadataHiderSettingTab extends PluginSettingTab {
 
 		let noteEl = containerEl.createEl("p", {
 			text: {
-				en: `Note: the last metadata property will always be visible, in order to ensure the normal usage of "Add property".`,
-				zh: `注意：最后一个文档属性将始终可见，以确保正常使用“添加文档属性”。`,
-				"zh-TW": `注意：最後一個文档屬性將始終可見，以確保正常使用「添加文档屬性」。`,
+				en: `When the metadata properties table is focused, (i.e. inputting metadata properties), all metadata properties will be displayed, except "Metadata properties always to hide".`,
+				zh: `当文档属性（元数据）表格获得焦点时（即输入元数据），除“永远隐藏的文档属性”外的所有文档属性都将显示。`,
+				"zh-TW": `當文檔屬性（元數據）表格獲得焦點時（即輸入元數據），除「永遠隱藏的文件屬性」外的所有文檔屬性都將顯示。`,
 			}[lang] as string
-		}); //, even if it is hidden by the above settings.
+		});
 		noteEl.setAttribute("style", "color: gray; font-style: italic; margin-top: 30px;")
 	}
 }